+++ modules/theme-support/tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use Liberu\Foundation\Theme\Tests\TestCase;
+use Liberu\PackageTestbench\PackageTestCase;
 
-pest()->extend(TestCase::class)->in('Architecture', 'Integration', 'Unit');
+pest()->extend(PackageTestCase::class)->in('Unit');